refactor reference props
