+++ packages/tinacms/src/toolkit/fields/components/reference/model/reference-props.ts
@@ -1,0 +1,9 @@
+import { ReferenceFieldProps, Option } from './reference-field-props'
+
+export interface ReferenceProps {
+  name: string
+  input: any
+  field: ReferenceFieldProps
+  disabled?: boolean
+  options?: (Option | string)[]
+}
